feat: non packed moduleTypes now 256 bit to save on padding gas

+++ src/DataTypes.sol
@@ -115,7 +115,7 @@ function attestationDataRefEq(
 
 type PackedModuleTypes is uint32;
 
-type ModuleType is uint32;
+type ModuleType is uint256;
 
 using { moduleTypeEq as == } for ModuleType global;
 using { moduleTypeNeq as != } for ModuleType global;
